+++ semana19/projeto/src/services/idGenerator.ts
@@ -1,0 +1,5 @@
+import { v4 } from "uuid";
+
+const generateId = () => v4()
+
+export default generateId
